requirements.txt
absl-py==2.0.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.1
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cycler==0.10.0
debugpy==1.7.0
decorator==5.1.1
dill==0.3.7
entrypoints==0.4
Flask==2.2.5
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.8.0
idna==2.10
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==4.12.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mediapipe==0.9.0.1
nest-asyncio==1.5.8
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==3.11.0
psutil==5.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.6
rsa==4.9
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
supervision==0.11.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==2.0.7
wcwidth==0.2.8
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
